// web/src/api/generated/models/WishDetail.ts
export interface WishDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof WishDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof WishDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof WishDetail
     */
    createdBy: string;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof WishDetail
     */
    creator: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof WishDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof WishDetail
     */
    spaceId: string;
    /**
     * 
     * @type {WishStatus}
     * @memberof WishDetail
     */
    status: WishStatus;
    /**
     * 
     * @type {string}
     * @memberof WishDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof WishDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof WishDetail
     */
    version: number;
}